refactor: useTheme Hook
