+++ components/Navbar/Navbar.module.scss
@@ -87,9 +87,9 @@
     }
 
     img {
-      @include imageSize(40px, 40px, contain);
+      @include imageSize(45px, 45px, contain);
       @include borderWithRadius(2px solid #9c27b0, 50%);
-      padding: 5px;
+      padding: 6px;
       cursor: pointer;
     }
   }
